// barq-system-frontend/src/sdk/types.gen.ts
export type ArtificialIntelligenceBulletsEntity = {
  id: number;
  text: string;
  icon_id: number;
  artificial_intelligence_id: number;
  icon: MediaEntity;
  artificial_intelligence_bullets_id_artificial_intelligence_bullets_translations: Array<ArtificialIntelligenceBulletsTranslationsEntity>;
  artificial_intelligence_bullets_parent: ArtificialIntelligenceEntity;
  created_at: string;
  updated_at: string;
};

export type ArtificialIntelligenceBulletsTranslationsEntity = {
  id: number;
  artificial_intelligence_bullets_id: number;
  artificial_intelligence_bullets_translation_artificial_intelligence_bullets?: ArtificialIntelligenceBulletsEntity | null;
  text: string;
  language: 'ar' | 'en';
};

export type ArtificialIntelligenceEntity = {
  id: number;
  sub_headline: string;
  image_id: number;
  logo_id: number;
  image: MediaEntity;
  logo: MediaEntity;
  artificial_intelligence_id_artificial_intelligence_translations: Array<ArtificialIntelligenceTranslationsEntity>;
  artificial_intelligence_bullets_id_artificial_intelligence_bullets: Array<ArtificialIntelligenceBulletsEntity>;
  created_at: string;
  updated_at: string;
};

export type ArtificialIntelligenceTranslationsEntity = {
  id: number;
  artificial_intelligence_id: number;
  artificial_intelligence_translation_artificial_intelligence?: ArtificialIntelligenceEntity | null;
  sub_headline: string;
  language: 'ar' | 'en';
};

export type AutomationHeroEntity = {
  id: number;
  title: string;
  description: string;
  hero_id: number;
  hero: MediaEntity;
  logos: Array<MediaEntity>;
  automation_hero_id_automation_hero_translations: Array<AutomationHeroTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type AutomationHeroTranslationsEntity = {
  id: number;
  automation_hero_id: number;
  automation_hero_translation_automation_hero?: AutomationHeroEntity | null;
  title: string;
  description: string;
  language: 'ar' | 'en';
};
// ...
export type BusinessAutomationBulletsEntity = {
  id: number;
  text: string;
  icon_id: number;
  business_automation_id: number;
  icon: MediaEntity;
  business_automation_bullets_id_business_automation_bullets_translations: Array<BusinessAutomationBulletsTranslationsEntity>;
  business_automation_bullets_parent: BusinessAutomationEntity;
  created_at: string;
  updated_at: string;
};

export type BusinessAutomationBulletsTranslationsEntity = {
  id: number;
  business_automation_bullets_id: number;
  business_automation_bullets_translation_business_automation_bullets?: BusinessAutomationBulletsEntity | null;
  text: string;
  language: 'ar' | 'en';
};

export type BusinessAutomationEntity = {
  id: number;
  sub_headline: string;
  image_id: number;
  logo_id: number;
  image: MediaEntity;
  logo: MediaEntity;
  business_automation_id_business_automation_translations: Array<BusinessAutomationTranslationsEntity>;
  business_automation_bullets_id_business_automation_bullets: Array<BusinessAutomationBulletsEntity>;
  created_at: string;
  updated_at: string;
};

export type BusinessAutomationTranslationsEntity = {
  id: number;
  business_automation_id: number;
  business_automation_translation_business_automation?: BusinessAutomationEntity | null;
  sub_headline: string;
  language: 'ar' | 'en';
};
// ...
export type CloudBulletsEntity = {
  id: number;
  text: string;
  icon_id: number;
  cloud_id: number;
  icon: MediaEntity;
  cloud_bullets_id_cloud_bullets_translations: Array<CloudBulletsTranslationsEntity>;
  cloud_bullets_parent: CloudEntity;
  created_at: string;
  updated_at: string;
};

export type CloudBulletsTranslationsEntity = {
  id: number;
  cloud_bullets_id: number;
  cloud_bullets_translation_cloud_bullets?: CloudBulletsEntity | null;
  text: string;
  language: 'ar' | 'en';
};

export type CloudEntity = {
  id: number;
  sub_headline: string;
  image_id: number;
  logo_id: number;
  logo: MediaEntity;
  cloud_id_cloud_translations: Array<CloudTranslationsEntity>;
  cloud_bullets_id_cloud_bullets: Array<CloudBulletsEntity>;
  created_at: string;
  updated_at: string;
};

export type CloudTranslationsEntity = {
  id: number;
  cloud_id: number;
  cloud_translation_cloud?: CloudEntity | null;
  sub_headline: string;
  language: 'ar' | 'en';
};
// ...
export type type = 'soc' | 'cybersecurity' | 'grc';
// ...
export type DataManagementBulletsEntity = {
  id: number;
  text: string;
  icon_id: number;
  data_management_id: number;
  icon: MediaEntity;
  data_management_bullets_id_data_management_bullets_translations: Array<DataManagementBulletsTranslationsEntity>;
  data_management_bullets_parent: DataManagementEntity;
  created_at: string;
  updated_at: string;
};

export type DataManagementBulletsTranslationsEntity = {
  id: number;
  data_management_bullets_id: number;
  data_management_bullets_translation_data_management_bullets?: DataManagementBulletsEntity | null;
  text: string;
  language: 'ar' | 'en';
};

export type DataManagementEntity = {
  id: number;
  sub_headline: string;
  image_id: number;
  logo_id: number;
  image: MediaEntity;
  logo: MediaEntity;
  data_management_id_data_management_translations: Array<DataManagementTranslationsEntity>;
  data_management_bullets_id_data_management_bullets: Array<DataManagementBulletsEntity>;
  created_at: string;
  updated_at: string;
};

export type DataManagementTranslationsEntity = {
  id: number;
  data_management_id: number;
  data_management_translation_data_management?: DataManagementEntity | null;
  sub_headline: string;
  language: 'ar' | 'en';
};
// ...
export type id = 'ASC' | 'DESC';

export type title = 'ASC' | 'DESC';

export type description = 'ASC' | 'DESC';

export type image_id = 'ASC' | 'DESC';

export type created_at = 'ASC' | 'DESC';

export type updated_at = 'ASC' | 'DESC';
// ...
export type logo_id = 'asc' | 'desc';
// ...
export type sub_headline = 'asc' | 'desc';
// ...
export type hero_id = 'asc' | 'desc';
// ...
export type number = 'asc' | 'desc';
// ...
export type ArtificialIntelligenceControllerReadResponse = {
  data?: Array<ArtificialIntelligenceEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type AutomationHeroControllerReadResponse = {
  data?: Array<AutomationHeroEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type BusinessAutomationControllerReadResponse = {
  data?: Array<BusinessAutomationEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type DataManagementControllerReadResponse = {
  data?: Array<DataManagementEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type CloudSectionControllerReadResponse = {
  data?: Array<CloudEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};